- Summary: [Accessibility]Title Format Dialog, press "Tab" button, can not focus on "Outline" block
- Bugzilla Bug (s) Resolved:
144159, 144160, 143685

- Description:
The reason why outline visible button can't get focus by tab is that default control's enabled state is false. In swt, default enabled state is inherited from parent. As a custom control, LineAttributesComposite extends Composite, but overrided isEnabled() method. Hence, a new button created under this composite won't get focus because of disabled state inherited from the parent.
Solution: not override isEnabled() method in custom control

- Tests Description: Manual test
- Notes to Build Team: None
- Notes to Developers: None
- Notes to QA: None
- Notes to Documentation: None

- Files Edited:

- Files Added:

- Files Deleted

+++ chart/org.eclipse.birt.chart.ui.extension/src/org/eclipse/birt/chart/ui/swt/composites/IconCanvas.java
@@ -159,7 +159,7 @@ public class IconCanvas extends Canvas
 
 		if ( horizontal != null )
 		{
-			horizontal.setIncrement( (int) ( getClientArea( ).width / 100 ) );
+			horizontal.setIncrement( ( getClientArea( ).width / 100 ) );
 			horizontal.setPageIncrement( getClientArea( ).width );
 
 			if ( imageBound.width * sx > cw )
@@ -175,13 +175,13 @@ public class IconCanvas extends Canvas
 				tx = ( cw - imageBound.width * sx ) / 2;
 			}
 			horizontal.setSelection( (int) ( -tx ) );
-			horizontal.setThumb( (int) ( getClientArea( ).width ) );
+			horizontal.setThumb( ( getClientArea( ).width ) );
 		}
 		ScrollBar vertical = getVerticalBar( );
 		if ( vertical != null )
 		{
-			vertical.setIncrement( (int) ( getClientArea( ).height / 100 ) );
-			vertical.setPageIncrement( (int) ( getClientArea( ).height ) );
+			vertical.setIncrement( ( getClientArea( ).height / 100 ) );
+			vertical.setPageIncrement( ( getClientArea( ).height ) );
 			if ( imageBound.height * sy > ch )
 			{
 				vertical.setMaximum( (int) ( imageBound.height * sy ) );
@@ -195,7 +195,7 @@ public class IconCanvas extends Canvas
 				ty = ( ch - imageBound.height * sy ) / 2;
 			}
 			vertical.setSelection( (int) ( -ty ) );
-			vertical.setThumb( (int) ( getClientArea( ).height ) );
+			vertical.setThumb( ( getClientArea( ).height ) );
 		}
 
 		af = AffineTransform.getScaleInstance( sx, sy );
@@ -210,7 +210,6 @@ public class IconCanvas extends Canvas
 	 * 
 	 * @param filename
 	 * 
-	 * @return
 	 */
 	public Image loadImage( String filename )
 	{
@@ -237,7 +236,6 @@ public class IconCanvas extends Canvas
 	 * Load the image from a URL.
 	 * 
 	 * @param url
-	 * @return
 	 */
 	public Image loadImage( URL url )
 	{
@@ -266,7 +264,6 @@ public class IconCanvas extends Canvas
 	 * 
 	 * @param filename
 	 * 
-	 * @return
 	 */
 	public Image loadImage( InputStream is )
 	{
